requirements.txt
fastapi>=0.110
uvicorn
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
xgboost>=2.0
tensorflow>=2.15
joblib>=1.3
psycopg2-binary>=2.9
APScheduler>=3.10
python-dotenv>=1.0
pydantic-settings>=2.4
